#7: upgrade Spring Boot to 2.1.0 and Spring to 5.1.2.

+++ webui/qav-server/src/main/java/de/qaware/qav/server/controller/GraphController.java
@@ -170,7 +170,7 @@ public class GraphController {
 
         resultList.sort(Comparator.comparing(AbstractGraphElement::getName));
 
-        int startIndex = pageable.getOffset();
+        int startIndex = (int) pageable.getOffset();
         if (startIndex > resultList.size()) {
             startIndex = resultList.size();
         }
